+++ packages/plugin-packages/mashroom-portal-demo-alternative-theme/src/plugin/dev_server.ts
@@ -1,26 +1,28 @@
-// @flow
 /* eslint no-console: off */
 
 import express from 'express';
-import exphbs from 'express-handlebars';
-import helpers from './handlebar_helpers';
-
-import type {ExpressRequest, ExpressResponse} from '@mashroom/mashroom/type-definitions';
+import {createEngine} from 'express-react-views';
+import themeParams from './theme_params';
+import type {Request, Response} from 'express';
 import type {MashroomPortalPageRenderModel} from '@mashroom/mashroom-portal/type-definitions';
+
+themeParams.setParams({
+    mashroomVersion: '1.0.0',
+});
 
 const app = express();
 
 app.use('/resources', express.static('dist/public'));
 
-const hbs = exphbs.create({
-    helpers,
-    defaultLayout: '',
+const engine = createEngine({
+    transformViews: false
 });
 
-app.engine('handlebars', hbs.engine);
-app.set('view engine', 'handlebars');
+app.engine('js', engine);
+app.set('view engine', 'js');
+app.set('views', `${__dirname}/../views`);
 
-app.get('/', (req: ExpressRequest, res: ExpressResponse) => {
+app.get('/', (req: Request, res: Response) => {
 
     const model: MashroomPortalPageRenderModel = {
         portalName: 'Test Portal',
@@ -229,10 +231,6 @@ app.get('/', (req: ExpressRequest, res: ExpressResponse) => {
     res.render('portal', model);
 });
 
-app.listen(5066, (error: ?Error) => {
-    if (error) {
-        console.error(error);
-        return;
-    }
+app.listen(5066, () => {
     console.info('Server started at port 5066');
 });