+++ packages/plugin-packages/mashroom-portal-demo-alternative-theme/src/plugin/modules.d.ts
@@ -1,0 +1,6 @@
+
+declare module 'express-react-views' {
+    export const createEngine: (settings: {
+           transformViews: boolean;
+        }) => any;
+}