+++ packages/plugin-packages/mashroom-portal-demo-alternative-theme/src/plugin/theme_params.ts
@@ -1,0 +1,17 @@
+
+type ThemeParams = {
+    mashroomVersion: string;
+}
+
+let params: ThemeParams = {
+    mashroomVersion: '',
+};
+
+export default {
+    get mashroomVersion() {
+        return params.mashroomVersion;
+    },
+    setParams(_params: ThemeParams) {
+        params = _params;
+    },
+};